Move device management routes to own module
