+++ kolibri/plugins/device_management/assets/src/routes/wizardTransitionRoutes.js
@@ -3,10 +3,10 @@ import store from 'kolibri.coreVue.vuex.store';
 import {
   showAvailableChannelsPage,
   showSelectContentPage,
-} from './state/actions/contentWizardActions';
-import { updateTreeViewTopic } from './state/actions/selectContentActions';
-import { ContentWizardPages } from './constants';
-import { selectContentTopicLink } from './views/manage-content-page/manageContentLinks';
+} from '../state/actions/contentWizardActions';
+import { updateTreeViewTopic } from '../state/actions/selectContentActions';
+import { ContentWizardPages } from '../constants';
+import { selectContentTopicLink } from '../views/manage-content-page/manageContentLinks';
 
 // To update the treeview topic programatically
 export function navigateToTopicUrl(node) {
